Feat: 내 명함 삭제 이미지 로직 구현

+++ src/main/java/WELLET/welletServer/card/service/MyCardService.java
@@ -93,8 +93,8 @@ public class MyCardService {
         Card card = cardRepository.findByOwnerId(memberId)
                 .orElseThrow(() -> new CardException(CardErrorCode.CARD_NOT_FOUND));
         cardRepository.delete(card);
-        CardImage cardImage = cardImageRepository.findByCard(card);
-        cardImageRepository.delete(cardImage);
-        s3FileUploader.deleteFile(card.getProfImgUrl(), "profile_image");
+        if (card.getProfImgUrl() != null) {
+            s3FileUploader.deleteFile(card.getProfImgUrl(), "profile_image");
+        }
     }
 }
